lolrscrape/requirements.txt
beautifulsoup4==4.6.1
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
idna==2.7
lolrscrape==0.0.1
numpy==1.14.3
pandas==0.23.3
python-dateutil==2.7.3
python-memo==0.1.0
pytz==2018.5
requests==2.19.1
six==1.11.0
tenacity==4.12.0
urllib3==1.23
